refactor: rename dca as contact_distance
