+++ pNeuma_simulator/contact_distance/__init__.py
@@ -1,0 +1,3 @@
+from .contact_distance import calc_dtc, ellipses  # noqa F401
+
+__version__ = "0.0.0"
